refactor: api helpers
